Added SPI mock with transaction based expect
